% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{cluster_logging}
\alias{cluster_logging}
\title{Set up logging for running on the cluster.}
\usage{
cluster_logging(level_name = "info", error_file = NULL)
}
\arguments{
\item{level_name}{One of the strings (trace, debug, info, warn, error, fatal)
Default value is "info".}

\item{error_file}{The name of the file. If NULL,
    the error file will be named with the current date and time.

This sets logging so that warnings and above go to a file that is
named by the date and time, but all else goes to standard out.
It uses hierarchical loggers to control what goes to a file and what
goes to the console. Logging levels for warn, error, and fatal go
to a logging namespace called package-name.err, and that is directed to
a file.}
}
\description{
Set up logging for running on the cluster.
}
