% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\docType{data}
\name{.baseLogger}
\alias{.baseLogger}
\title{This is the namespace in which all logging messages
are placed, so that they can be controlled as a group.}
\format{
An object of class \code{character} of length 1.
}
\usage{
.baseLogger
}
\description{
This is the namespace in which all logging messages
are placed, so that they can be controlled as a group.
}
\keyword{datasets}
