% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\docType{data}
\name{.errorLogger}
\alias{.errorLogger}
\title{Create a child namespace that also logs errors to a file.}
\format{
An object of class \code{character} of length 1.
}
\usage{
.errorLogger
}
\description{
Create a child namespace that also logs errors to a file.
}
\keyword{datasets}
