% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_spy.R
\name{get.input.files}
\alias{get.input.files}
\title{Get the list of files that were recorded as inputs.}
\usage{
get.input.files(clear = FALSE)
}
\arguments{
\item{clear}{True in order to clear the list of input files.}
}
\value{
A list with one entry for each file.
}
\description{
Get the list of files that were recorded as inputs.
}
