% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_spy.R
\name{get.metadata}
\alias{get.metadata}
\title{Returns metadata for a file}
\usage{
get.metadata(path)
}
\arguments{
\item{path}{The path to the file from which to get file info.}
}
\value{
a list of file info containing creation time, last modified,
    and its path.
}
\description{
Returns metadata for a file
}
