% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_spy.R
\name{get.output.files}
\alias{get.output.files}
\title{Get the list of files that were recorded as outputs}
\usage{
get.output.files(clear = FALSE)
}
\arguments{
\item{clear}{True in order to clear the list of output files.}
}
\value{
A list with one entry for each file.
}
\description{
Get the list of files that were recorded as outputs
}
