% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_module}
\alias{log_module}
\title{Set the logging level on some part of a package.}
\usage{
log_module(module, level_name = "debug")
}
\arguments{
\item{module}{The string name of a section of the package.}

\item{level_name}{One of the strings (trace, debug, info, warn, error,
    fatal). The default is "debug".

If you want to debug one module, then you can set its level differently
here from the rest of the package. If the package is named `macro`
and this function is called with `log_module("mosquito", "debug")`, then
this will set two loggers to debug level: `macro.mosquito` and
`macro.err.mosquito`, where the second one logs warnings and above to a file.}
}
\description{
Set the logging level on some part of a package.
}
