% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_spy.R
\name{prov.input.file}
\alias{prov.input.file}
\title{Record that this code used this input file.}
\usage{
prov.input.file(input, role)
}
\arguments{
\item{input}{The filename of the input file.}

\item{role}{A string to identify what this file does for this program.}
}
\description{
Record that this code used this input file.
}
