% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_spy.R
\name{prov.output.file}
\alias{prov.output.file}
\title{Record that this code used this output file.}
\usage{
prov.output.file(output, role)
}
\arguments{
\item{output}{The filename of the output file.}

\item{role}{A string to identify what this file does for this program.}
}
\description{
Record that this code used this output file.
}
