% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_spy.R
\name{write.meta.data}
\alias{write.meta.data}
\title{Write the list of input files to a file on disk.}
\usage{
write.meta.data(path)
}
\arguments{
\item{path}{The file path to which to write the information.}
}
\description{
Write the list of input files to a file on disk.
}
